pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.116.0",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.4",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
